# static/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: Mechanic Workshop API
  description: API documentation for Customers, Mechanics, Service Tickets, and Inventory.
# API server configuration specifying host, base path, and HTTP scheme
host: localhost:5000
basePath: /api
schemes:
  - http
# Security definition for JWT Bearer token authentication via Authorization header
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: >
      JWT Token authentication. Use 'Bearer <your-token>' format.
# Tags categorizing API endpoints by resource type
...
paths:
...
  /inventory:
    post:
      tags:
        - Inventory
      summary: Create Part
      description: "Creates a new inventory part."
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/InventoryInput'
            example:
              name: "Brake Pad"
              sku: "BP-12345"
              description: "Front brake pads for sedan models"
              price: 49.99
      responses:
        201:
          description: "Inventory part created successfully"
          schema:
            $ref: '#/definitions/Inventory'
          examples:
            application/json:
              id: 101
              name: "Brake Pad"
              sku: "BP-12345"
              description: "Front brake pads for sedan models"
              price: 49.99
# Retrieve a paginated list of all inventory parts (requires authentication)
  /inventory/:
    get:
      tags:
        - Inventory
      summary: Find All Parts (Paginated)
      description: Returns a paginated list of inventory parts.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Results per page (default: 10)"
      responses:
        200:
          description: A paginated list of inventory parts.
          schema:
            type: object
            properties:
              inventory:
                type: array
                items:
                  $ref: '#/definitions/Inventory'
              total:
                type: integer
              pages:
                type: integer
              current_page:
                type: integer
# Retrieve inventory part details by ID (requires authentication)
  /inventory/{id}:
    get:
      tags:
        - Inventory
      summary: Find Part by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Inventory part details
          schema:
            $ref: '#/definitions/Inventory'
# Update an inventory part's details by ID (requires authentication)
    put:
      tags:
        - Inventory
      summary: Update Part by ID
      description: "Updates an existing inventory part."
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: ID of the inventory part to update
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/InventoryInput'
            example:
              name: "Brake Pad"
              sku: "BP-12345"
              description: "Front brake pads for sedan models (updated)"
              price: 54.99
      responses:
        200:
          description: "Inventory part updated successfully"
          schema:
            $ref: '#/definitions/Inventory'
          examples:
            application/json:
              id: 101
              name: "Brake Pad"
              sku: "BP-12345"
              description: "Front brake pads for sedan models (updated)"
              price: 54.99
# Delete an inventory part by ID (requires authentication)
    delete:
      tags:
        - Inventory
      summary: Delete Part by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: Inventory part deleted successfully
# Customer object model representing a registered customer’s details
definitions:
...
  Inventory:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      sku:
        type: string
      description:
        type: string
      price:
        type: number
    required:
      - name
      - sku
      - description
      - price
# Input model for creating or updating an inventory part with required fields
  InventoryInput:
    type: object
    required:
      - name
      - sku
      - description
      - price
    properties:
      name:
        type: string
      sku:
        type: string
      description:
        type: string
      price:
        type: number